refactor: remove non utilized code or import
