+++ components/OptionCard.tsx
@@ -1,11 +1,5 @@
 import React from 'react';
-import {
-  StyleSheet,
-  TouchableOpacity,
-  TouchableOpacityProps,
-  View,
-  ViewStyle,
-} from 'react-native';
+import {StyleSheet, TouchableOpacity, View, ViewStyle} from 'react-native';
 import {Icon, useTheme} from '@tsejerome/ui-kitten-components';
 
 import Text from './Text';
